refactor: split spClient.ts God File into sp/ modules (1629 to 1294 lines, -21%)
